requirements.txt
# Core dependencies for FWI Resolution Enhancement

# Data handling
numpy>=1.21.0
pandas>=1.3.0
xarray>=0.19.0
netCDF4>=1.5.7

# Machine Learning
scikit-learn>=1.0.0
scipy>=1.7.0

# Deep Learning (optional, for Method 4)
# torch>=1.10.0
# torchvision>=0.11.0

# Climate data
cdsapi>=0.7.4

# Visualization
matplotlib>=3.4.0

# Utilities
pyyaml>=5.4.0
tqdm>=4.62.0

# Development
jupyter>=1.0.0
ipykernel>=6.0.0
pytest>=6.2.0